% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_collectors_name.R
\name{get_collectors_name}
\alias{get_collectors_name}
\title{Simplify and extract the name of collectors}
\usage{
get_collectors_name(x = "", min_char = 2)
}
\arguments{
\item{x}{name of the collectors}

\item{min_char}{min characters in name}
}
\value{
Standard name of the main collector
}
\description{
This function simplifies and extracts the name of one collector
}
\examples{
\donttest{
# Normal conditions
get_collectors_name('Anderberg, Arne & Anderberg, Anna-Lena')
get_collectors_name('Schlegel, Henrik Adolf Leonard & Arnell, Hampus Wilrshelm')
get_collectors_name('Tomitarô-Makino, John')

# Edge cases
get_collectors_name(NA)
get_collectors_name('')
get_collectors_name('et al.')

}


}
