% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_GBIF_archive.R
\name{import_GBIF_archive}
\alias{import_GBIF_archive}
\title{(Step 1) Import the GBIF occurrence Darwin Core Archive data}
\usage{
import_GBIF_archive(gbif_occurrence_file = "", only_PRESERVED_SPECIMEN = F)
}
\arguments{
\item{gbif_occurrence_file}{GBIF occurrence Darwin Core Archive data. See details for more information.}

\item{only_PRESERVED_SPECIMEN}{if TRUE, occurrence data will be filtered by \code{Ctrl_basisOfRecord="PRESERVED_SPECIMEN"}}
}
\value{
A list with duration and 3 data.table: "occ" for processed occurrence data,"occ_gbif_issue" for checked GBIF issues and "summary" for import summary.
}
\description{
Returns a list contains processed GBIF occurrence data and GBIF issue data which is useful for downstream analysis.
}
\details{
gbif_occurrence_file might be a path to your Darwin Core standard file which is downloaded from GBIF.
The Darwin Core Archive (DwC-A) is a compact package (a ZIP file) contains interconnected text files and enables data publishers to share their data using a common terminology.
gbif_occurrence_file could also be a path to "occurrence.txt" which is decompressed from your your Darwin Core standard file.
}
\examples{
\donttest{

help(import_GBIF_archive)


}
}
