% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_richness.R
\name{plot_richness}
\alias{plot_richness}
\title{Plot a richness map for native refined records in 1 degree grids}
\usage{
plot_richness(
  records_refined = NA,
  main = "richness plot",
  xlab = "Longitude",
  ylab = "Latitude"
)
}
\arguments{
\item{records_refined}{the refined occurrence list from \code{coordinate_refine}}

\item{main}{tittle of the plot}

\item{xlab}{label of x axis of the plot}

\item{ylab}{label of y axis of the plot}
}
\value{
a list include a presence table and a richness S4 object
}
\description{
This optional module plots a richness map of native refined records on grids. It
is based on two functions \code{lets.presab.points} and \code{plot.PresenceAbsence} from R package \code{letsR}
(see references below for more information), but fully leverages vectorization techniques
to avoid looping when filling large matrices, thus achieving nearly a hundredfold speedup.
}
\examples{
\dontrun{
plot_richness(records_refined = refine, main = 'richness map',
 xlab = "Longitude", ylab = "Latitude")
}
}
\references{
Vilela, B. and Villalobos, F. (2015), letsR: a new R package for data handling and analysis in macroecology. Methods Ecol Evol, 6: 1229-1234. https://doi.org/10.1111/2041-210X.12401
}
