% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_collectors_dictionary.R
\name{prepare_collectors_dictionary}
\alias{prepare_collectors_dictionary}
\title{Prepare a list with surname of the main collector}
\usage{
prepare_collectors_dictionary(
  occ_import = NA,
  surname_selection_type = "largest_string",
  max_words_name = 6,
  min_characters_in_name = 4
)
}
\arguments{
\item{occ_import}{imported GBIF records}

\item{surname_selection_type}{allows you to select any of two types of the last name:

\strong{\code{large_string}} = word with the largest number of characters.

\strong{\code{last_name}} = literally the last name of the main collector, with more than two characters.}

\item{max_words_name}{assist in checking the length of name}

\item{min_characters_in_name}{min characters in name}
}
\value{
A list with duration and 2 data.table: "my_dictionary" for your processed
collectors dictionary and "ref_dictionary" for collectors name reference dictionary
}
\description{
It returns a list with the last name of main collector following the unique key \code{recordedBy}.

Consistent recording of the primary collectors and their last name is thus essential, and a \strong{collector dictionary}
is provided for this purpose. The \code{prepare_collectors_dictionary} function is utilized to extract the last
name of the primary collector from the \code{recordedBy} field and to compile a list linking the last name
of the primary collector with the raw data in \code{recordedBy}

It is advisable to verify the \strong{last name} of the primary collector in the \code{nameRecordedBy_Standard}
field. The objective is to standardize the last name of the primary collector, which is automatically
derived from the \code{recordedBy} field. This is accomplished by standardizing the text string to begin with
an uppercase character and replacing non-ASCII characters, ensuring that the primary collector associated
with a collection event is consistently represented by the same character string.

Once processed, the dictionary can be checked in the future.
}
\details{
If \code{recordedBy} is present in the ref_dictionary, the function returns the last name of
the primary collector associated with the specific \code{recordedBy} key. Otherwise, it
returns the last name of the primary collector, extracted from the \code{recordedBy} field.

It is recommended to curate the last name of the primary collector, which is automatically
derived from the \code{recordedBy} field. The goal is to standardize the last name of the primary
collector. The primary botanical collector of a sample is consistently identified by the
same last name, standardized in capital letters with non-ASCII characters replaced.
}
\examples{
\donttest{
help(collectors_prepare_dictionary)
}
}
