% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ref_dictionary}
\alias{ref_dictionary}
\title{Reference dictionary for names of collectors}
\format{
A data.table with 115702 rows and 2 columns
\describe{
\item{Ctrl_recordedBy}{original collectors name}
\item{Ctrl_nameRecordedBy_Standard}{processed collectors name}
}
}
\source{
\itemize{
\item \href{https://github.com/pablopains/parseGBIF/tree/main/collectorDictionary}{collectorDictionary}
}
}
\usage{
ref_dictionary
}
\description{
A distilled reference dictionary for names of collectors
}
\references{
de Melo, P.H.A., Bystriakova, N., Lucas, E. et al. A new R package to parse plant species occurrence records into unique collection events efficiently reduces data redundancy. Sci Rep 14, 5450 (2024). https://doi.org/10.1038/s41598-024-56158-3
}
\keyword{datasets}
