% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ref_wcvp_names}
\alias{ref_wcvp_names}
\title{The World Checklist of Vascular Plants: names}
\format{
A data.table with 443,347 rows and 2 variables:
\describe{
\item{plant_name_id}{World Checklist of Vascular Plants (WCVP) identifier}
\item{taxon name}{Concatenation of genus with species and, where applicable, infraspecific epithets to make a binomial or trinomial name.}
}
}
\source{
\url{http://sftp.kew.org/pub/data-repositories/WCVP/wcvp.zip}
}
\usage{
ref_wcvp_names
}
\description{
A reduced dataset containing the taxa names from the WCVP that have distributions information
}
\references{
Govaerts, R., Nic Lughadha, E., Black, N. et al. The World Checklist of Vascular Plants, a continuously updated resource for exploring global plant diversity. Sci Data 8, 215 (2021). https://doi.org/10.1038/s41597-021-00997-6
}
\keyword{datasets}
