% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refine_records.R
\name{refine_records}
\alias{refine_records}
\title{Restore duplicate records from voucher, validate coordinates and extract their World Geographical Scheme for Recording Plant Distributions}
\usage{
refine_records(
  voucher = NA,
  threads = 4,
  save_path = NA,
  tests = c("capitals", "centroids", "equal", "gbif", "institutions", "outliers", "seas",
    "zeros")
)
}
\arguments{
\item{voucher}{UltraGBIF_voucher from \code{\link{set_digital_voucher}}}

\item{threads}{threads requirement, a positive real number, default is 4}

\item{save_path}{the local path where you want to save the final result}

\item{tests}{CoordinateCleaner checks. Choose one or more from \code{c("capitals","centroids","equal","gbif","institutions","outliers","seas","zeros")}}
}
\value{
UltraGBIF_refine list with duration and 2 data.table: \code{all_records} for all refined records,
\code{native_records} for native records of them.
}
\description{
Restore key usable information for vouchers by consolidating data from duplicate records belonging to identical collection events, validate coordinates and extract their World Geographical Scheme for Recording Plant Distributions.
}
\details{
If set \code{save_path}, \code{usable_refined_records.csv.gz} that includes usable records and
\code{native_refined_records.csv.gz} that includes usable records will be write to the path.
They are the main products of UltraGBIF.
}
\examples{
\dontrun{
refine <- refine_records(voucher = voucher,
threads = 4,
save_path = getwd(),
tests = c("capitals", "centroids", "equal", "gbif",
"institutions", "outliers", "seas","zeros"))
}
}
\seealso{
\code{\link[CoordinateCleaner]{clean_coordinates}}
}
