% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seas_ref}
\alias{seas_ref}
\title{A simple features object of the world land map}
\format{
A simple features object
\describe{
\item{featurecla}{Land area}
\item{geometry}{Geometry unit}
}
}
\source{
\code{rnaturalearth::ne_download(scale = 110,type = 'land',category = 'physical',returnclass = "sf")}
}
\usage{
seas_ref
}
\description{
An integrated land map
}
\keyword{datasets}
