% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usecores.R
\name{usecores}
\alias{usecores}
\title{Detect and optimize threads requirement}
\usage{
usecores(x)
}
\arguments{
\item{x}{Threads requirement, a positive real number, default is 4}
}
\value{
A positive real number of treads
}
\description{
Input a positive number and return optimized treads requirement.
}
\details{
If a positive integer is input,
return a positive integer not exceeding the total number of threads.
If a non-integer positive number is input, automatically calculate and return a reasonable number of threads.
}
\examples{
\dontrun{
usecores(4)
usecores(60)
usecores(0.45)
}

}
