% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wcvp_check_name.R
\name{wcvp_check_name}
\alias{wcvp_check_name}
\title{(Copilot) Check taxon names by WCVP database}
\usage{
wcvp_check_name(
  searchedName = "Tephrosia lasiochlaena Cowie",
  wcvp_names_database = NA,
  if_author_fails_try_without_combinations = TRUE
)
}
\arguments{
\item{searchedName}{scientific name, with or without author}

\item{wcvp_names_database}{processed wcvp database}

\item{if_author_fails_try_without_combinations}{option for partial verification of the authorship of the species. Remove the authors of combinations, in parentheses}
}
\value{
data.table with WCVP fields
}
\description{
Use the \href{https://powo.science.kew.org//}{World Checklist of Vascular Plants}
\href{https://github.com/matildabrown/rWCVPdata}{database} to check accepted names and update synonyms.
}
\details{
About the \href{https://powo.science.kew.org/about-wcvp}{World Checklist of Vascular Plants}
The \code{searchNotes} field assigns values based on the following conditions:
\itemize{
\item \strong{Accepted}: When a single authorless scientific name exists in the \code{TAXON_name} list with \code{TAXON_STATUS = "Accepted"}, \code{verified_speciesName = 100}.
\item \strong{Accepted among homonyms}: When multiple authorless scientific names are present in the \code{TAXON_name} list, but only one homonym has \code{TAXON_STATUS = "Accepted"}, \verb{verified_speciesName = number of matches / 100}.
\item \strong{Homonyms}: When multiple authorless scientific names appear in the \code{TAXON_name} list and either multiple or none of the homonyms have \code{TAXON_STATUS = "Accepted"}, \verb{verified_speciesName = number of matches / 100}. Prior to homonym detection, attempts to match authorless scientific names in \code{TAXON_name} with authors in \code{TAXON_AUTHORS} failed, resulting in \code{verified_author = 0}.
\item \strong{Not Found}: When the authorless scientific name is absent from the \code{TAXON_name} list.
\item \strong{Unplaced}: When a single authorless scientific name exists in the \code{TAXON_name} list with \code{TAXON_STATUS = "Unplaced"}.
\item \strong{Updated}: When a single authorless scientific name is present in the \code{TAXON_name} list, and both \code{ACCEPTED_PLANT_NAME_ID} and the species ID differ (i.e., \code{ACCEPTED_PLANT_NAME_ID != plant_name_id_of_searchedName}), the fields \code{taxon_status_of_searchedName}, \code{plant_name_id_of_searchedName}, and \code{taxon_authors_of_searchedName} are populated as follows:
\item For \code{searchNotes = "Updated"}: These fields record data from the originally queried scientific name.
\item For \code{searchNotes = "Homonyms"}: These fields record data from homonymous synonyms, separated by "|".
}

The \code{verified_author} field is assigned as follows:
\itemize{
\item \strong{100}: When a direct match exists between the authorless scientific name in \code{TAXON_name} and an author in \code{TAXON_AUTHORS}.
\item \strong{50}: When a partial correspondence exists between the authorless scientific name in \code{TAXON_name} and an author in \code{TAXON_AUTHORS}, without combination.
\item \strong{0}: When no author is present in \code{TAXON_AUTHORS}, irrespective of correspondence with the authorless scientific name in \code{TAXON_name}.
}
}
\examples{
\donttest{
require(data.table)
## It might take a little time
wcvp_names_database = setDT(rWCVPdata::wcvp_names)[,`:=`(TAXON_NAME_U=toupper(taxon_name),
TAXON_AUTHORS_U=toupper(taxon_authors))]

# 1) Updated
wcvp_check_name(searchedName = 'Hemistylus brasiliensis Wedd.',
               wcvp_names_database = wcvp_names_database,
               if_author_fails_try_without_combinations = TRUE)

# 2) Accepted
wcvp_check_name(searchedName = 'Hemistylus boehmerioides Wedd. ex Warm.',
               wcvp_names_database = wcvp_names_database,
               if_author_fails_try_without_combinations = TRUE)

# 3) Unplaced - taxon_status = Unplaced
wcvp_check_name(searchedName = 'Leucosyke australis Unruh',
               wcvp_names_database = wcvp_names_database,
               if_author_fails_try_without_combinations = TRUE)

# 4) Accepted among homonyms - When author is not informed. In this case, one of
# the homonyms, taxon_status is accepted
wcvp_check_name(searchedName = 'Parietaria cretica',
               wcvp_names_database = wcvp_names_database,
               if_author_fails_try_without_combinations = TRUE)

# When author is informed
wcvp_check_name(searchedName = 'Parietaria cretica L.',
               wcvp_names_database = wcvp_names_database,
               if_author_fails_try_without_combinations = TRUE)

# When author is informed
wcvp_check_name(searchedName = 'Parietaria cretica Moris',
               wcvp_names_database = wcvp_names_database,
               if_author_fails_try_without_combinations = TRUE)

# 5) Homonyms - When author is not informed. In this case, none of the homonyms,
# taxon_status is Accepted
wcvp_check_name(searchedName = 'Laportea peltata',
               wcvp_names_database = wcvp_names_database,
               if_author_fails_try_without_combinations = TRUE)

# When author is informed
wcvp_check_name(searchedName = 'Laportea peltata Gaudich. & Decne.',
               wcvp_names_database = wcvp_names_database,
               if_author_fails_try_without_combinations = TRUE)

# When author is informed
wcvp_check_name(searchedName = 'Laportea peltata (Blume) Gaudich.',
               wcvp_names_database = wcvp_names_database,
               if_author_fails_try_without_combinations = TRUE)
}

}
