% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_occ_name.R
\name{check_occ_name}
\alias{check_occ_name}
\title{Using integrated Taxonomic Name Resolution Service to check taxon names and update synonyms under World Checklist of Vascular Plants}
\usage{
check_occ_name(occ_import = NA, accuracy = 0.9)
}
\arguments{
\item{occ_import}{imported GBIF records}

\item{accuracy}{numeric. If specified, only matches with a score greater than or equal to the supplied accuracy level will be returned. If left, the default threshold will be 0.9.}
}
\value{
A list with duration and two data.table: "summary" with taxa list and "occ_wcvp_check_name" with WCVP fields
}
\description{
Names of taxa are checked using integrated Taxonomic Name Resolution Service (TNRS)
}
\details{
\itemize{
\item \href{https://tnrs.biendata.org/about/}{About TNRS}
\item \href{https://powo.science.kew.org//}{World Checklist of Vascular Plants}
}
}
\examples{
\dontrun{
taxa_checked <- check_occ_name(occ_import = occ_import,accuracy = 0.9)
}
}
\seealso{
\code{\link[TNRS]{TNRS}}
}
