% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_collection_mark.R
\encoding{UTF-8}
\name{generate_collection_mark}
\alias{generate_collection_mark}
\title{Generate unique collection event}
\usage{
generate_collection_mark(occ_import = NA, dictionary = NA)
}
\arguments{
\item{occ_import}{imported GBIF records}

\item{dictionary}{your processed dictionary from \code{prepare_collectors_dictionary}}
}
\value{
A list with duration and 3 data.table:
"occ_collectorsDictionary" for update result fields only, "summary" for summary and
"CollectorsDictionary_add" for new collectors that can be added to the collector dictionary
that can be reused in the future.
}
\description{
It creates a key to identify the physical and digital duplicates of a given collection event.
It combines the last name of primary collector with the collector number and the botanical family (family +
recordByStandardized + recordNumber_Standard) that groups the duplicates of the same unique collection event.
It also identifies new collectors to be added to the collector dictionary and that can be reused in the future.

Include \strong{\code{recordedByStandardized}} field with verified last name of the primary collector.

Include \strong{\code{recordNumber_Standard}} field with only numbers from \strong{\code{recordNumber}}.

Create the collection event key to group duplicates in the \strong{\code{key_family_recordedBy_recordNumber}} field
following the fields:

\strong{\code{family} + \code{recordedByStandardized} + \code{recordNumber_Standard}}.
}
\details{
Fields created for each incident record:

\code{nameRecordedBy_Standard}

\code{recordNumber_Standard}

\code{key_family_recordedBy_recordNumber}

\code{key_year_recordedBy_recordNumber}

#' A critical step in parsing duplicate records involves generating a robust key for each unique collecting
event (i.e., \emph{gathering}) to facilitate duplicate recognition. To achieve this, a string is created by
concatenating the \strong{taxon family}, the \strong{last name of the primary collector}, and the \strong{collection number}.
}
\examples{
\donttest{
help(generate_collection_mark)
}

}
