% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_records.R
\name{import_records}
\alias{import_records}
\title{Import GBIF occurrence records}
\usage{
import_records(GBIF_file = "", only_PRESERVED_SPECIMEN = F)
}
\arguments{
\item{GBIF_file}{path to the downloaded zip file from GBIF}

\item{only_PRESERVED_SPECIMEN}{if TRUE, occurrence records are filtered by \code{basisOfRecord="PRESERVED_SPECIMEN"}}
}
\value{
UltraGBIF_import list with duration and 3 data.table: "occ" for processed occurrence data,"occ_gbif_issue" for checked GBIF issues and "summary" for import summary.
}
\description{
Load initial GBIF records and extract issues for downstream analysis.
}
\details{
This step is the beginning of UltraGBIF progress, which makes downloaded file into initial GBIF records.
}
\examples{
\dontrun{
## run as example for Saxifraga occurrence records
library(UltraGBIF)
url <- 'https://api.gbif.org/v1/occurrence/download/request/0021523-250402121839773.zip'
gbif_occurrence_file <- paste0(getwd(),'/Saxfriga.zip')
curl::curl_download(url,gbif_occurrence_file,quiet = F) ## cost time and total are 68692523 bytes
occ_import <- import_records(GBIF_file = gbif_occurrence_file,
                            only_PRESERVED_SPECIMEN = T)
head(occ_import$summary,5)
}
}
\seealso{
\itemize{
\item \code{\link{unzip}} for automatically decompressing
\item Saxifraga occurrence records: \doi{https://doi.org/10.15468/dl.bythb4}
}
}
