% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ref_dictionary}
\alias{ref_dictionary}
\title{ref_dictionary}
\format{
A data.table with 115702 rows and 2 columns
\describe{
\item{Ctrl_recordedBy}{original collectors name}
\item{Ctrl_nameRecordedBy_Standard}{processed collectors name}
}
}
\source{
\itemize{
\item \href{https://github.com/pablopains/parseGBIF/tree/main/collectorDictionary}{collectorDictionary}
}
}
\usage{
ref_dictionary
}
\description{
ref_dictionary
}
\keyword{internal}
