% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{seas_ref}
\alias{seas_ref}
\title{seas_ref}
\format{
A simple features object
\describe{
\item{featurecla}{Land area}
\item{geometry}{Geometry unit}
}
}
\source{
rnaturalearth::ne_download(scale = 110,type = 'land',category = 'physical',returnclass = "sf")
}
\usage{
seas_ref
}
\description{
seas_ref
}
\keyword{internal}
