% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usecores.R
\name{usecores}
\alias{usecores}
\title{(Copilot) Detect threads and optimize your threads requirement}
\usage{
usecores(x)
}
\arguments{
\item{x}{your threads requirement, a number >0.}
}
\value{
A number of treads
}
\description{
Input a positive number and return optimized treads requirement.
}
\details{
If a positive integer is input,
return a positive integer not exceeding the total number of threads.
If a non-integer positive number is input, automatically calculate and return a reasonable number of threads.
}
\examples{
\donttest{
usecores(4)
usecores(1/3)}

}
